Great Hyperspace War
Created a div to hold the description and vertical-aligned it to the bottom.

Made its background a linear gradient from transparent to gray and back to transparent to give the appearance of them fading away.

Gave the text wider left and right margins so it would stay on the visible part of the gradient.

Wrote a description of the Great Hyperspace War.

Set to display: none to hide it until it is unhidden by JavaScript.