Restructure divs
Removed the .holdSpace divs around each character.

Created a separate div next to each character and set it to display: none.

Moved all the transformation IDs into this new div and changed them to classes.

.charSum is only present in the first div and .charDescrip is only present in the second div to make different text appear.

Used JavaScript IDs to set the first div to visibility: hidden and the second div to display: block when clicked on.

This fixes a glitch where the .holdSpace divs expanded to accommodate the grid layout but the .char divs inside them did not.