Slideshow HTML
Restructured the HTML layout of the slideshow.

Created the .eventCont div to hold this entire section.

Created four .primEvent divs to hold each one of the large images.

Numbered each one of the .primEvent divs.

Put a paragraph with a description of the image in each .primEvent div.

Added in the .prev and .next buttons.

Created the .secondEvent div to hold all the smaller images at the bottom.

Created four .eventCont divs inside .secondEvent div.

Put an .event div inside each .eventCont div.

Set the background image of each .event div.

Added a paragraph inside each .eventCont div with a label for the image.